requirements.txt
openai>=1.3.7
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.23.0
python-dotenv>=1.0.0
simpleaudio>=1.0.4 